* Don't rely on BMLog's stored normals, instead flag pbvh nodes to   update normals on undo/redo.

+++ source/blender/bmesh/intern/bmesh_log.h
@@ -115,13 +115,13 @@ void BM_log_before_all_removed(BMesh *bm, BMLog *log);
 const float *BM_log_original_vert_co(BMLog *log, BMVert *v);
 
 /* Get the logged normal of a vertex */
-const short *BM_log_original_vert_no(BMLog *log, BMVert *v);
+const float *BM_log_original_vert_no(BMLog *log, BMVert *v);
 
 /* Get the logged mask of a vertex */
 float BM_log_original_mask(BMLog *log, BMVert *v);
 
 /* Get the logged data of a vertex (avoid multiple lookups) */
-void BM_log_original_vert_data(BMLog *log, BMVert *v, const float **r_co, const short **r_no);
+void BM_log_original_vert_data(BMLog *log, BMVert *v, const float **r_co, const float **r_no);
 
 /* For internal use only (unit testing) */
 BMLogEntry *BM_log_current_entry(BMLog *log);
